+++ src/auth.js
@@ -1,6 +1,6 @@
 import NextAuth from "next-auth";
 import Credentials from "next-auth/providers/credentials";
-import { mFetch } from "@/util/MFetch";
+import { mFetch } from "@/util/mFetch";
 import { cookies } from "next/headers";
 
 export const { handlers, signIn, signOut, auth } = NextAuth({
